Moved Utility and Dev commands to cogs
